+++ src/main/java/com/mageddo/dnsproxyserver/docker/ContainerFacade.java
@@ -1,5 +1,6 @@
 package com.mageddo.dnsproxyserver.docker;
 
+import com.github.dockerjava.api.command.InspectContainerResponse;
 import com.github.dockerjava.api.model.Container;
 
 import java.util.List;
@@ -9,4 +10,8 @@ public interface ContainerFacade {
   List<Container> findNetworkContainers(String networkId);
 
   Container findById(String containerId);
+
+  List<Container> findActiveContainers();
+
+  InspectContainerResponse inspect(String id);
 }
