+++ src/main/java/com/mageddo/dnsproxyserver/docker/ContainerFacadeDefault.java
@@ -1,6 +1,7 @@
 package com.mageddo.dnsproxyserver.docker;
 
 import com.github.dockerjava.api.DockerClient;
+import com.github.dockerjava.api.command.InspectContainerResponse;
 import com.github.dockerjava.api.model.Container;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -37,4 +38,19 @@ public class ContainerFacadeDefault implements ContainerFacade {
       ;
   }
 
+  @Override
+  public List<Container> findActiveContainers() {
+    final var activeContainers = this.dockerClient
+      .listContainersCmd()
+      .withStatusFilter(Containers.RUNNING_STATUS_LIST)
+      .withLimit(1024)
+//      .withNetworkFilter()
+      .exec();
+    return activeContainers;
+  }
+
+  @Override
+  public InspectContainerResponse inspect(String id) {
+    return this.dockerClient.inspectContainerCmd(id).exec();
+  }
 }
