+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/docker/dataprovider/ContainerDAODefault.java
@@ -3,7 +3,6 @@ package com.mageddo.dnsproxyserver.server.dns.solver.docker.dataprovider;
 import com.github.dockerjava.api.DockerClient;
 import com.mageddo.dnsproxyserver.docker.ContainerFacade;
 import com.mageddo.dnsproxyserver.docker.Containers;
-import com.mageddo.dnsproxyserver.docker.DockerFacade;
 import com.mageddo.dnsproxyserver.server.dns.solver.docker.Container;
 import com.mageddo.dnsproxyserver.server.dns.solver.docker.dataprovider.mapper.ContainerMapper;
 import lombok.RequiredArgsConstructor;
@@ -21,7 +20,6 @@ import java.util.Collections;
 public class ContainerDAODefault implements ContainerDAO {
 
   private final DockerClient dockerClient;
-  private final DockerFacade dockerFacade;
   private final ContainerFacade containerFacade;
 
   public Container findDPSContainer() {
@@ -40,7 +38,7 @@ public class ContainerDAODefault implements ContainerDAO {
     return containers
       .stream()
       .findFirst()
-      .map(it -> this.dockerFacade.inspect(it.getId()))
+      .map(it -> this.containerFacade.inspect(it.getId()))
       .map(ContainerMapper::of)
       .orElse(null);
   }
